requirements.txt
streamlit
transformers
torch
torchvision
torchaudio
pillow
accelerate